feat: updata spring ai dependency to 1.0.0-M3

+++ spring-ai-alibaba-core/src/main/java/com/alibaba/cloud/ai/autoconfigure/dashscope/DashScopeChatProperties.java
@@ -45,7 +45,7 @@ public class DashScopeChatProperties extends DashScopeParentProperties {
 	/**
 	 * Default temperature speed.
 	 */
-	private static final Float DEFAULT_TEMPERATURE = 0.8f;
+	private static final Double DEFAULT_TEMPERATURE = 0.8d;
 
 	/**
 	 * Enable Dashscope ai chat client.
